pyproject.toml
[tool.poetry]
name = "jax_ensemble"
version = "0.1.0"
description = "A simple neural network ensemble in JAX."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^1.2.0"
flax = "^0.6.0"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
